refactor: moved test stubs to appropriate files
