+++ src/providers/google/GoogleProvider.test.ts
@@ -472,3 +472,42 @@ describe('GoogleProvider conference mapping', () => {
     );
   });
 });
+
+describe('GoogleProvider deleted recurring instances', () => {
+  it('requests cancelled instances from Google Calendar', async () => {
+    const plugin = {
+      app: {
+        vault: {
+          getAbstractFileByPath: jest.fn()
+        },
+        metadataCache: {
+          getFileCache: jest.fn(),
+          on: jest.fn(),
+          offref: jest.fn()
+        }
+      }
+    } as unknown as FullCalendarPlugin;
+    const provider = new GoogleProvider(
+      {
+        id: 'google_1',
+        name: 'Google Calendar',
+        calendarId: 'primary'
+      },
+      plugin
+    );
+
+    jest.spyOn(provider['authManager'], 'getTokenForSource').mockResolvedValue('token');
+    PluginState.getSettings = jest.fn().mockReturnValue({ displayTimezone: 'UTC' });
+    const requestMock = jest.mocked(makeAuthenticatedRequest);
+    requestMock.mockResolvedValue({ items: [] });
+
+    await provider.getEvents({
+      start: new Date('2026-08-01T00:00:00Z'),
+      end: new Date('2026-09-01T00:00:00Z')
+    });
+
+    const requestUrl = requestMock.mock.calls[0]?.[1];
+    expect(requestUrl).toBeDefined();
+    expect(new URL(requestUrl || '').searchParams.get('showDeleted')).toBe('true');
+  });
+});
